+++ .github/workflows/build.yml
@@ -18,6 +18,6 @@ jobs:
 
     - name: test
       run: |
-        curl -Lo mkn https://github.com/mkn/mkn/releases/download/latest/mkn_nix
+        curl -fL --retry 3 --retry-delay 2 -o mkn https://github.com/mkn/mkn/releases/download/latest/mkn_nix
         chmod +x mkn
         KLOG=3 ./mkn clean build run -dtKOgp cpu -a "-std=c++20" test -W 9
