+++ .gitignore
@@ -1,4 +1,4 @@
 bin
 .mkn
 *.cui
-.clangd
+*.clangd
