+++ inc/mkn/gpu/any/def.hpp
@@ -1,0 +1,4 @@
+#ifndef _MKN_GPU_ANY_DEF_HPP_
+#define _MKN_GPU_ANY_DEF_HPP_
+
+#endif /*_MKN_GPU_ANY_DEF_HPP_*/
