--- inc/mkn/gpu/cli.hpp
@@ -1,55 +1,0 @@
-/**
-Copyright (c) 2024, Philip Deegan.
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are
-met:
-
-    * Redistributions of source code must retain the above copyright
-notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above
-copyright notice, this list of conditions and the following disclaimer
-in the documentation and/or other materials provided with the
-distribution.
-    * Neither the name of Philip Deegan nor the names of its
-contributors may be used to endorse or promote products derived from
-this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-// IWYU pragma: private, include "mkn/gpu.hpp"
-#ifndef _MKN_GPU_CLI_HPP_
-#define _MKN_GPU_CLI_HPP_
-
-#include "mkn/kul/env.hpp"
-#include "mkn/kul/string.hpp"
-
-namespace mkn::gpu {
-
-template <typename Device>
-struct Cli {
-  constexpr static inline char const* MKN_GPU_BX_THREADS = "MKN_GPU_BX_THREADS";
-
-  auto bx_threads() const {
-    if (kul::env::EXISTS(MKN_GPU_BX_THREADS))
-      return kul::String::INT32(kul::env::GET(MKN_GPU_BX_THREADS));
-    return dev.maxThreadsPerBlock;
-  }
-
-  Device const& dev;
-};
-
-} /* namespace mkn::gpu */
-
-#endif /*_MKN_GPU_CLI_HPP_*/
