+++ inc/mkn/gpu/cpu/api.hpp
@@ -1,0 +1,4 @@
+#ifndef _MKN_GPU_CPU_API_HPP_
+#define _MKN_GPU_CPU_API_HPP_
+
+#endif /*_MKN_GPU_CPU_API_HPP_*/
