+++ inc/mkn/gpu/cpu/cls.hpp
@@ -1,0 +1,4 @@
+#ifndef _MKN_GPU_CPU_CLS_HPP_
+#define _MKN_GPU_CPU_CLS_HPP_
+
+#endif /*_MKN_GPU_CPU_CLS_HPP_*/
