+++ inc/mkn/gpu/cpu/def.hpp
@@ -1,0 +1,4 @@
+#ifndef _MKN_GPU_CPU_DEF_HPP_
+#define _MKN_GPU_CPU_DEF_HPP_
+
+#endif /*_MKN_GPU_CPU_DEF_HPP_*/
