+++ inc/mkn/gpu/def.hpp
@@ -3,9 +3,46 @@
 
 #include <type_traits>
 
+#if !defined(MKN_GPU_ROCM) and __has_include("hip/hip_runtime.h")
+#define MKN_GPU_ROCM 1
+#endif
+#if !defined(MKN_GPU_ROCM)
+#define MKN_GPU_ROCM 0
+#endif
+
+#if !defined(MKN_GPU_CUDA) and __has_include(<cuda_runtime.h>)
+#define MKN_GPU_CUDA 1
+#endif
+#if !defined(MKN_GPU_CUDA)
+#define MKN_GPU_CUDA 0
+#endif
+
+#if MKN_GPU_CUDA && MKN_GPU_ROCM && !defined(MKN_GPU_FN_PER_NS)
+#define MKN_GPU_FN_PER_NS 1
+#endif
+
+#if !defined(MKN_GPU_FN_PER_NS)
+#define MKN_GPU_FN_PER_NS 0
+#endif
+
+#if MKN_GPU_CUDA == 0 && MKN_GPU_ROCM == 0 && !defined(MKN_GPU_CPU)
+#define MKN_GPU_CPU 1
+#endif
+
+#if !defined(MKN_GPU_CPU)
+#define MKN_GPU_CPU 0
+#endif
+
 namespace mkn::gpu {
 
-#if defined(MKN_GPU_CPU)
+struct CompileFlags {
+  bool constexpr static withCUDA = MKN_GPU_CUDA;
+  bool constexpr static withROCM = MKN_GPU_ROCM;
+  bool constexpr static withCPU = MKN_GPU_CPU;
+  bool constexpr static perNamespace = MKN_GPU_FN_PER_NS;
+};
+
+#if MKN_GPU_CPU
 
 template <typename T>
 static constexpr bool is_floating_point_v = std::is_floating_point_v<T>;
