--- inc/mkn/gpu/devfunc.hpp
@@ -1,44 +1,0 @@
-// IWYU pragma: private, include "mkn/gpu.hpp"
-/**
-Copyright (c) 2025, Philip Deegan.
-All rights reserved.
-
-Redistribution and use in source and binary forms, with or without
-modification, are permitted provided that the following conditions are
-met:
-
-    * Redistributions of source code must retain the above copyright
-notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above
-copyright notice, this list of conditions and the following disclaimer
-in the documentation and/or other materials provided with the
-distribution.
-    * Neither the name of Philip Deegan nor the names of its
-contributors may be used to endorse or promote products derived from
-this software without specific prior written permission.
-
-THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-"AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
-SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-#ifndef _MKN_GPU_DEVFUNC_HPP_
-#define _MKN_GPU_DEVFUNC_HPP_
-
-template <bool sync = true, typename T>
-void __device__ zero(T* const t, std::size_t const size) {
-  std::size_t chunk = 0;
-  auto const each = size / warpSize;
-  for (; chunk < each; ++chunk) t[chunk * warpSize + threadIdx.x] = 0;
-  if (threadIdx.x < size - (warpSize * each)) t[chunk * warpSize + threadIdx.x] = 0;
-  if constexpr (sync) __syncthreads();
-}
-
-#endif /* _MKN_GPU_DEVFUNC_HPP_ */
