+++ inc/mkn/gpu/stream.hpp
