+++ inc/mkn/gpu/tuple.hpp
@@ -31,7 +31,6 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #ifndef _MKN_GPU_TUPLE_HPP_
 #define _MKN_GPU_TUPLE_HPP_
 
-// #include "mkn/gpu.hpp"
 #include "mkn/kul/tuple.hpp"
 
 namespace mkn::gpu {
